requirements/common.txt

# django
Django==2.2.8
django-su==0.8.0
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
djangorestframework-camel-case==1.0.3
django-cors-headers==3.1.0
python-dotenv==0.10.3

# posgres
psycopg2==2.8.3
psycopg2-binary==2.8.3
pillow >= 6.1.0

# Celery
celery[redis]==4.2.1

pytz==2019.1
sqlparse==0.3.0

Markdown==3.1.1

#Swagger
drf-yasg==1.16.0

#django-storages
boto3==1.9.223
django-storages==1.7.1
